% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpvalue.R
\name{bootpval}
\alias{bootpval}
\title{Title}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{The sample}

\item{conf.level}{The value of 1 - alpha for your confidence interval used in getting the p-value}

\item{iter}{The number of iterations in the bootstrap method}

\item{mu0}{The null hypothesis value for mu}

\item{test}{Which type of test you want to use}
}
\value{
A list of the important statistics used in bootstrapping the p-value.
}
\description{
Title
}
