% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math4753kouryRpackage.R
\docType{package}
\name{math4753kouryRPackage}
\alias{math4753kouryRPackage}
\title{Koury Package}
\description{
We are going to put a lot of stuff here in this class.
}
\author{
Will Koury \email{wkoury@ou.edu}
}
