% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Title MyBoot}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{population vector}

\item{fun}{function you want to execute on the sample vector}

\item{alpha}{confidence interval alpha value}

\item{...}{any other parameters you want to have when the histogram is generated}
}
\value{
nothing
}
\description{
Title MyBoot
}
\examples{
myboot(x=(1:9999),fun="median")
}
