% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNCurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{a}{from negative infinity to a}
}
\value{
nothing
}
\description{
Title
}
\examples{
myncurve(0,1,0.5)
}
