% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlot.R
\name{myplot}
\alias{myplot}
\title{Some curve that Dr. Stewart gave us..}
\usage{
myplot(x = 1:4)
}
\arguments{
\item{x}{the input to the function}
}
\value{
the output of the function
}
\description{
Some curve that Dr. Stewart gave us..
}
\examples{
myplot(1:5)
myplot(5.34)
}
