% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCLTu.R
\name{packagecltu}
\alias{packagecltu}
\title{Title}
\usage{
packagecltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Number n of each sample}

\item{iter}{The number of iterations}

\item{a}{Lower bound of the uniform distribution}

\item{b}{Upper bound of the uniform distribution}
}
\value{

}
\description{
Title
}
