% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruceFunction.R
\name{plotSpruce}
\alias{plotSpruce}
\title{Plot some data from the SPRUCE data set.}
\usage{
plotSpruce(filename)
}
\arguments{
\item{filename}{the name of the spruce data set file}
}
\value{
Nothing
}
\description{
Plot some data from the SPRUCE data set.
}
\examples{
plotSpruce()
}
